pyproject.toml
[tool.poetry]
name = "pysonet"
version = "0.1.0"
description = "Social network by Django"
authors = ["DJWOMS <socanime@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
djangorestframework = "^3.11.1"
pillow = "^7.2.0"
psycopg2 = "^2.8.5"
djoser = "^2.0.3"
djangorestframework-simplejwt = "*"
drf-yasg = "^1.17.1"
django-cors-headers = "^3.5.0"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
